+++ NEWS.md
@@ -27,6 +27,7 @@
 * New option to configure soft wrapping for R Markdown files, and command to change the soft wrap mode of the editor on the fly (#2341)
 * Add option `www-url-path-prefix` to force a path on auth cookies (Pro #1608)
 * New Command Palette for searching and running build-in commands and add-ins (#5168)
+* Option to display Console error and message output in same color as regular output (#7029)
 
 ### RStudio Server Pro